Simple plugin to expose dsl to app instances
Dancer2 routes pass the app instance as their first param into the route
coderef. You can do a lot with that; but there are DSL methods that are
not (easily) callable.

This plugin exposes the DSL object to the app instance. YAGNI applies.
However, I'm writing code generators based on a spec (OpenAPI) and the
controllers live in some namespace that doesn't become a Dancer2 app.